requirements.txt
absl-py==1.2.0
astor==0.8.1
certifi==2022.9.14
charset-normalizer==2.1.1
docopt==0.6.2
gast==0.5.3
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.48.1
h5py==3.7.0
idna==3.4
importlib-metadata==4.12.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
networkx==2.6.3
numpy==1.21.6
Pillow==9.2.0
pipreqs==0.4.11
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-louvain==0.16
pytz==2022.2.1
PyYAML==6.0
rdflib==6.2.0
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==1.14.0
tensorboardX==2.5.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==2.0.1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.1
pandas==1.3.5
biopython